refactor(workspace): refactor components for easier code navigation
